Fix of issue #XD-748, #XD-874, #XD-875, #XD-867 (#30)
All files are divided into segments with sizes equal to the size of the cache page. Those segments are also called pages.

The following metadata are stored at the end of each page:
1. The offset of the start of the first loggable inside the page.
2. xxHash code of the page's content

That allows not only the detection of broken content but also restoring the rest of the content in future releases.
The consistency of the pages is checked during the following cases:
1. The page is read from the disk. This check could be switched off by option "exodus.checkPagesAtRuntime"
2. The backup routine is examining the consistency of all pages.
3. If the database is incorrectly closed, an assessment of whole pages is automatically triggered.

The system automatically switches into read-only mode if the database detects that one or more pages are broken.

The concept of startup metadata was implemented. That is a chunk of metadata which contains information about:
1. Maximum size of a single file segment.
2. Size of the cache page.
3. The version of the currently used binary format of database files.
4. The version of the binary format of the metadata file itself.
5. The address of the root of the metadata tree.
6. Flag whether the database was correctly closed.

The implementation of the shadow copy achieves the durability of the startup metadata.

So as of this change, it is not allowed, or better to say that such changes will be automatically reverted, to change the size of the page in the cache or the maximum size of the file after creating the database.

The introduction of metadata inside each page implies overhead of calculation of offset of the physical position of different parts of the loggable inside the page. To avoid such overhead additional precautions were taken to ensure that all loggable fit inside a single page, if possible. Such changes mean that, in the worst case, data will consume 5% more space on disk, but in practice, space overhead should be less than the worst-case scenario.

The newly introduced format is incompatible with the previous one, so during the start of the database, it is triggered routine which automatically migrates all the data.

This change requires JDK 11+ to be used.