Issue #XD-904 fixed
When using the file system for storing blobs, writing the blob's content is performed
after committing metadata in the Xodus Environment. That implies a delay between transaction acknowledgement
and the actual writing of data which can cause a situation when blob content can be incorrectly read in
another thread. To avoid such situations blob content is checked before returning to the user.
If the content is incomplete, the thread will wait a specified time in seconds until
the transaction wholly writes blob content. Otherwise, an exception will be thrown.