change the Quantizer.splitVectorsByPartitions interface
Add centroids to the return value. It lets test the method more thoroughly.
Add VectorReader to the parameters set. NorExplicitQuantizer does not do PQ k-means clustering (because it can not) and does a regular k-means clustering. For the regular k-means clustering the original vectors are necessary. So we have those original vectors in the interface now.