XD-1082 fix OSequenceImpl implementation
The previous implementation created a session on every access to the sequence (get, increment, etc.) The new implementation creates a new session only when both the sequence has not been created yet and the current session has an active transaction. Also, the new implementation makes sure that the sequence is created right away and not lazily.